+++ src/models/Customer.php
@@ -9,20 +9,24 @@ class Customer extends Model
 {
 
 	public $customerNumber;
+	public $vatZone;
+	public $paymentTerms;
+	public $name;
 
-	public static function transformFromOrder($order){
+	public static function transformFromOrder($order)
+	{
 		$billingAddress = $order->getBillingAddress();
 
 		//Check for business customer
 		$customerData = Economic::getInstance()->getCustomers()->getCustomerByVatNumber($billingAddress->businessTaxId);
 
 		//Check for private customer if no business
-		if(!$customerData){
+		if (!$customerData) {
 			$customerData = Economic::getInstance()->getCustomers()->getCustomerByEmail($order->email);
 		}
 
-		//TODO Customer doesn't exists in E-conomic, create new one
-		if(!$customerData){
+		//Customer not created, do it now
+		if (!$customerData) {
 			$customerData = Economic::getInstance()->getCustomers()->createCustomerFromOrder($order);
 		}
 
@@ -31,20 +35,70 @@ class Customer extends Model
 		return $customer;
 	}
 
-	public static function transform($object){
-
+	public static function transform($object)
+	{
 		$customer = new self();
-
 		$customer->setCustomerNumber($object->customerNumber);
-
+		$customer->setVatZone(VatZone::transform($object->vatZone));
+		$customer->setName($object->name);
+		$customer->setCustomerGroup(CustomerGroup::transform($object->customerGroup));
+		$customer->setPaymentTerms(PaymentTerms::transform($object->paymentTerms));
 		return $customer;
 	}
 
-	public function setCustomerNumber($value){
-
+	public function setCustomerNumber($value)
+	{
 		$this->customerNumber = $value;
-
 		return $this;
+	}
 
+	public function getCustomerNumber()
+	{
+		return $this->customerNumber;
+	}
+
+	public function setCustomerGroup(CustomerGroup $value)
+	{
+		$this->customerGroup = $value;
+		return $this;
+	}
+	public function getCustomerGroup()
+	{
+		return $this->customerGroup;
+	}
+
+	public function setName(string $value)
+	{
+		$this->name = $value;
+		return $this;
+	}
+	public function getName()
+	{
+		return $this->name;
+	}
+
+	public function setPaymentTerms(PaymentTerms $value)
+	{
+		$this->paymentTerms = $value;
+		return $this;
+	}
+	public function getPaymentTerms()
+	{
+		return $this->paymentTerms;
+	}
+
+	public function setVatZone(VatZone $value)
+	{
+		$this->vatZone = $value;
+		return $this;
+	}
+	public function getVatZone()
+	{
+		return $this->vatZone;
+	}
+
+	public function asArray()
+	{
+		return (array) $this;
 	}
 }