+++ src/models/Recipient.php
@@ -41,9 +41,17 @@ class Recipient extends Model
 		return $this;
 	}
 
+	public function getName(){
+		return $this->name;
+	}
+
 	public function setAddress($value){
 		$this->address = $value;
 		return $this;
+	}
+
+	public function getAddress(){
+		return $this->address;
 	}
 
 	public function setZip($value){
@@ -51,13 +59,25 @@ class Recipient extends Model
 		return $this;
 	}
 
+	public function getZip($value){
+		return $this->zip;
+	}
+
 	public function setCity($value){
 		$this->city = $value;
 		return $this;
 	}
 
+	public function getCity($value){
+		return $this->city;
+	}
+
 	public function setVatZone(VatZone $value){
 		$this->vatZone = $value;
 		return $this;
+	}
+
+	public function getVatZone(){
+		return $this->vatZone;
 	}
 }